Add packages, add string resources
