+++ app/src/main/java/com/example/gitsearch/ui/Navigation.kt
@@ -1,14 +1,21 @@
-package com.example.gitsearch
+package com.example.gitsearch.ui
 
 import android.util.Log
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.res.painterResource
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
+import com.example.gitsearch.R
+import com.example.gitsearch.ui.screen.AboutScreen
+import com.example.gitsearch.ui.screen.RepoDetail
+import com.example.gitsearch.ui.screen.SearchScreen
+import com.example.gitsearch.ui.screen.UserDetail
 
 @Composable
 fun Navigation(navController: NavHostController) {
     NavHost(navController = navController, startDestination = Screen.Search.route) {
+
         composable(Screen.Search.route) {
             SearchScreen(
                 onSearch = {
@@ -17,13 +24,20 @@ fun Navigation(navController: NavHostController) {
                     navController.navigate(Screen.UserDetail.route)
                 })
         }
+
         composable(Screen.UserDetail.route) {
-            UserDetail(onClickRepo = {
+            UserDetail(painterResource(id = R.drawable.ic_launcher_foreground), onClickRepo = {
                 navController.navigate(Screen.RepoDetail.route)
             })
         }
+
         composable(Screen.RepoDetail.route) {
             RepoDetail()
         }
+
+        composable(Screen.About.route) {
+            AboutScreen()
+        }
+
     }
 }
