+++ app/src/main/java/com/example/gitsearch/ui/Screen.kt
@@ -1,7 +1,8 @@
-package com.example.gitsearch
+package com.example.gitsearch.ui
 
 sealed class Screen(val route: String) {
     object Search : Screen("search")
     object UserDetail : Screen("userDetail")
     object RepoDetail : Screen("repoDetail")
+    object About : Screen("about")
 }
