+++ app/src/main/java/com/example/gitsearch/ui/screen/AboutScreen.kt
@@ -1,0 +1,8 @@
+package com.example.gitsearch.ui.screen
+
+import androidx.compose.runtime.Composable
+
+@Composable
+fun AboutScreen() {
+    TODO("About screen not implemented")
+}
